Update prompt in WeatherApiEndpoints to include location ID
The prompt string in the WeatherApiEndpoints class has been modified to include the location identifier (id). The new prompt now specifies the location by saying "Considering the following weather conditions, suggest an activity for today in {id}." This change aims to provide more context in the prompt by including the location.